Preparing a clean slate to begin the transition to CakePHP.
All the original source files have been taken out in this commit so that an initial version based around CakePHP can be dropped in even though the model itself is still in some flux. Using an MVC framework should speed development as well as make it easier to maintain in the future.