Initial commit of the CakePHP based framework
This skeleton should allow for rapid development of a self contained app that uses Cake's MVC model to separate the database handling, business logic, and user interface into distinct concerns. While the app itself lacks that much complexity it also serves as an opportunity to get more familiar with the framework in case it pays dividends with more complex projects down the road.