Creating a visitors table to start tying together the front end with the back end interface
This feels a bit shaky but should function fine until I have a better handle on CakePHP's internals. I am not even sure if the models are strictly needed due to some 'magic' within Cake but I want to follow Rails-like conventions as closely as possible. Consider the code at this point highly in flux.