Cleaning up Laravel structure in preparation for adding in the front end registration process
The first cut might just use a static workflow but hooked into the controller. From there wiring up the database in Eloquent should just be a matter of determining how to do validation, figure out how to write some basic tests, and then deploying it in production. Another step that needs to be taken is to document how to deploy using the Alias directive in Apache but that particular task is outside the scope of this repository.