Adding a static landing page
Need to clean this code up further by abstracting the layout from the rest. The good news is that from here the rest of the interface should fall neatly in to place one method at a time. The next step is to write the registration piece which includes the various views for each type of visitor