Can now choose a role type and continue to the terms of use
The controller workflow is now smart enough to respond to different types of roles and select the right form. The views themselves still need to finish
being customized as well as setting up a Request object that needs to be
smart enough to handle all the different types of objects. The solution
may be to create a number of FormRequest objects that subclass from one
master. All information should be stored locally and not written to the
database until terms of use have been accepted.