Adding validation rules for registration page
Based on the predetermined role, pick the validation rules that apply
to a particular form. Currently there are several bugs that still need
to be fixed such as the fact that it redirects all the way back to th
initial page instead of the form. Also it does not yet remember values
although this should be doable be making everything 'Laravel-able'