Fixed problem with selecting rules for which fields should be validated based on role
The code still does not behave properly and two outstanding bugs keep it
from being near ready to properly test. The first involves a strange error
that keeps kicking back to 'new' instead of reporting errors. The second
has to do with making the 'Go Back' button more intelligent so that it
is able to know which page to return to rather than always bouncing all
the way back to the start (albiet with prepopulated values)