Back button functionality now implemented for multipage forms
The entire submission process is now complete except that the details at
each point need to be written to the  object. Also when going
backwards it may be required to extract certain things in order for the
form to remain populated. This should be the next step before going back
to fine tune the error message formatting