Data now persists across submissions
By using @include rather than nest() the registration data will persist
across the entire session. The next step is to capture the signature,
write everything to the database, and then purge the session. It also
needs to be set up to prevent duplication based on email address with
some sort of error message. Default handling would be fine for now