Registrations now write to database
Registrations can now be saved to the database with optional fields
populated and form only items glossed over. At this point it would be
a great idea to spend a few hours writing tests to ensure that only
certain fields get populated for different views and that other workflows
behave as expected.

If everything passes muster the next step is to work on friendlier error
messages and the registration side of the workflow