Refactored registration migration
Since a couple of tasks need access to the old database tables they have
been refactored into a simple model. This can be abandoned in a future
version once the migrations are no longer required. In the meantime the
reporting and migration tasks should be able to leverage them to make
life easier