Enable live checkins against the user's name and authentication
It should now be possible to check in if a matching record is found in
the User table. The next phase should be to tie checkin to registration
somehow to enable people to register if they have problems. Also it should
be smart enough to not repeatedly check people in over and over again if
they have already checked in for the day. This can be done silently.